Add utests and move table creation to new
